segundo commit

+++ tarefas.py
@@ -3,7 +3,7 @@ numero = int (input("Informe um numero para saber o dobro dele -> "))
 print ("O dobro do seu numero é", numero * 2)
 
 #Valor da area do quadrado
-valor= int (input("Informe a medida do lado do quadrado -> "))
+valor= float (input("Informe a medida do lado do quadrado -> "))
 print ("A area do quadrado é", valor * valor)
 
 #Valor da idade em meses e dias
@@ -11,9 +11,9 @@ idade =int (input("Informe sua idade para calcularmos ela em Meses e dias -> "))
 print ("A sua idade é", idade, "Em Meses são" , idade * 12, "E", idade * 365,"Dias")
 
 #Media de tres numeros
-numeros1 = int(input("Me informe o seu primeiro numero -> "))
-numeros2 = int(input("Me informe o seu segundo numero -> "))
-numeros3 = int(input("Me informe o seu terceiro numero -> "))
-media = int (numeros1 + numeros2 + numeros3) / 3
+numeros1 = float(input("Me informe o seu primeiro numero -> "))
+numeros2 = float(input("Me informe o seu segundo numero -> "))
+numeros3 = float(input("Me informe o seu terceiro numero -> "))
+media = float (numeros1 + numeros2 + numeros3) / 3
 print ("A media de" , numeros1,",", numeros2, "e", numeros3, "é igual a", media)
